examples/chat-agent/pyproject.toml
[project]
name = "chat-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.3",
    "langchain-openai>=0.3",
    "langsmith>=0.2",
    "python-dotenv>=1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "langgraph-cli>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chat_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "smoke: always-on import and cold-start checks",
    "integration: live OpenAI-backed checks",
]
